+++ database/migrations/2022_18_05_000000_create_authentication_log_records_table.php
@@ -14,6 +14,7 @@ class CreateAuthenticationLogRecordsTable extends Migration
             $table->string('authenticatable_type')->nullable();
             $table->json('credentials')->nullable();
             $table->ipAddress('user_ip')->nullable();
+            $table->string('guard')->nullable();
             $table->string('eventType');
             $table->dateTime('recorded_at');
         });